Moves RelatedProducts components into components/relatedProducts folder:
